+++ packages/frontend/src/pages/scaling/summary/props/getScalingSummaryView.ts
@@ -1,8 +1,7 @@
-import { Layer2, layer2s, Layer3, safeGetTokenByAssetId } from '@l2beat/config'
+import { Layer2, layer2s, Layer3 } from '@l2beat/config'
 import {
   ImplementationChangeReportApiResponse,
   TvlApiResponse,
-  TvlApiToken,
   VerificationStatus,
 } from '@l2beat/shared-pure'
 
@@ -10,12 +9,8 @@ import { orderByTvl } from '../../../../utils/orderByTvl'
 import { getProjectTvlTooltipText } from '../../../../utils/project/getProjectTvlTooltipText'
 import { isAnySectionUnderReview } from '../../../../utils/project/isAnySectionUnderReview'
 import { getRiskValues } from '../../../../utils/risks/values'
-import { getTvlWarning } from '../../../../utils/tvl/getTVLBreakdown'
-import {
-  getTvlStats,
-  TvlStats,
-  unifyTokensResponse,
-} from '../../../../utils/tvl/getTvlStats'
+import { getTvlStats, TvlStats } from '../../../../utils/tvl/getTvlStats'
+import { getTvlWarnings } from '../../../../utils/tvl/getTVLWarnings'
 import { formatPercent, formatUSD } from '../../../../utils/utils'
 import { ScalingL2SummaryViewEntry, ScalingL3SummaryViewEntry } from '../types'
 import { ScalingSummaryViewProps } from '../view/ScalingSummaryView'
@@ -72,14 +67,7 @@ function getScalingL2SummaryEntry(
     stats = getTvlStats(apiProject, project.display.name, associatedTokens)
   }
 
-  const total = apiProject?.charts.hourly.data.at(-1)?.[1] ?? 0
-  const tokens = unifyTokensResponse(apiProject?.tokens)
-  const associatedRatio = getAssociatedRatio(associatedTokens, total, tokens)
-  const { warning, warningSeverity } = getTvlWarning(
-    associatedRatio,
-    project.display.name,
-    associatedTokens,
-  )
+  const tvlWarnings = getTvlWarnings(apiProject, associatedTokens, project)
 
   return {
     name: project.display.name,
@@ -95,15 +83,7 @@ function getScalingL2SummaryEntry(
     showProjectUnderReview: isAnySectionUnderReview(project),
     isUpcoming: project.isUpcoming,
     redWarning: project.display.redWarning,
-    tvlWarnings: [
-      warning && warningSeverity === 'bad'
-        ? {
-            content: warning,
-            sentiment: warningSeverity,
-          }
-        : undefined,
-      project.display.tvlWarning,
-    ],
+    tvlWarnings,
     tvl:
       stats && escrowsConfigured(project)
         ? {
@@ -160,20 +140,4 @@ function getScalingL3SummaryEntry(
 
 export function escrowsConfigured(project: Layer2) {
   return project.config.escrows.length > 0
-}
-
-function getAssociatedRatio(
-  associatedTokens: string[],
-  total: number,
-  tokens: TvlApiToken[],
-) {
-  let associated = 0
-
-  for (const token of tokens) {
-    const safeToken = safeGetTokenByAssetId(token.assetId)
-    if (safeToken && associatedTokens.includes(safeToken.symbol)) {
-      associated += token.usdValue
-    }
-  }
-  return total === 0 ? 0 : associated / total
 }
